Move common code from `build` and `run` to common base class
